add some changes

+++ jhash/src/jhash/FileSave.java
@@ -3,8 +3,6 @@ package jhash;
 
 import java.io.BufferedWriter;
 import java.io.File;
-import java.io.FileNotFoundException;
-import java.io.FileReader;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.util.Set;
@@ -13,12 +11,12 @@ import java.util.TreeMap;
 public class FileSave {
     
     
-     public int export_file (File file, TreeMap<Integer, String> treemap) throws  IOException{
+     public int export_file (File file, TreeMap<Long, String> treemap) throws  IOException{
          
          FileWriter fileReader = new FileWriter(file);
          BufferedWriter bufferedWriter = new BufferedWriter(fileReader);     
-        Set<Integer> keys = treemap.keySet();
-             for(Integer key: keys)
+        Set<Long> keys = treemap.keySet();
+             for(Long key: keys)
               {
                  bufferedWriter.write(treemap.get(key) + "\n");
               }
